大作业/第八组代码/requirements.txt
matplotlib==3.6.0
numpy==1.23.3
pandas==1.5.1
Pillow==9.3.0
scikit_learn==1.1.3
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
